unit09/stock-quant/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.18
aiosignal==1.3.2
altair==5.5.0
annotated-types==0.7.0
anyio==4.9.0
attrs==25.3.0
beautifulsoup4==4.13.4
blinker==1.9.0
bs4==0.0.2
cachetools==5.5.2
certifi==2025.4.26
charset-normalizer==3.4.1
click==8.1.8
dataclasses-json==0.6.7
distro==1.9.0
filelock==3.18.0
frozenlist==1.6.0
fsspec==2025.3.2
gitdb==4.0.12
GitPython==3.1.44
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.0
idna==3.10
Jinja2==3.1.6
jiter==0.9.0
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2025.4.1
langchain==0.3.24
langchain-community==0.3.23
langchain-core==0.3.56
langchain-experimental==0.3.4
langchain-ollama==0.2.2
langchain-openai==0.3.14
langchain-text-splitters==0.3.8
langchain-together==0.3.0
langgraph==0.4.0
langgraph-checkpoint==2.0.25
langgraph-prebuilt==0.1.8
langgraph-sdk==0.1.48
langsmith==0.1.147
llvmlite==0.44.0
lxml==5.4.0
MarkupSafe==3.0.2
marshmallow==3.26.1
mpmath==1.3.0
multidict==6.4.3
mypy_extensions==1.1.0
narwhals==1.37.1
networkx==3.4.2
numba==0.61.2
numpy==2.2.5
ollama==0.4.8
openai==1.76.2
orjson==3.10.18
ormsgpack==1.9.1
packaging==24.2
pandas==2.2.3
pandas-ta @ file:///Users/tranduythanh/BoxSync/genai-k3/unit09/stock-quant/pandas_ta-0.4.25b0-py3-none-any.whl#sha256=7cdb0b4f14a78d44f65771c6e6534f907665678972b7ddd041a284c6034f53c9
pillow==11.2.1
plotly==6.0.1
propcache==0.3.1
protobuf==6.30.2
pyarrow==20.0.0
pydantic==2.11.4
pydantic-settings==2.9.1
pydantic_core==2.33.2
pydeck==0.9.1
pypandoc==1.15
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2025.2
PyYAML==6.0.2
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
requests-toolbelt==1.0.0
rpds-py==0.24.0
setuptools==75.8.0
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
soupsieve==2.7
SQLAlchemy==2.0.40
sympy==1.14.0
tenacity==9.1.2
tiktoken==0.9.0
toml==0.10.2
torch==2.7.0
tornado==6.4.2
tqdm==4.67.1
typing-inspect==0.9.0
typing-inspection==0.4.0
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.4.0
-e git+ssh://git@github.com/phamdinhkhanh/vnquant.git@a98d48734890b13517c66acd3deba385580e4e53#egg=vnquant
wheel==0.45.1
wrapt==1.17.2
xxhash==3.5.0
yarl==1.20.0
